Split the UI into 2 pages:
- One for configuration
- One for selecting raw folders & doing the conversion
