+++ run_Waters2mzML.py
@@ -3,4 +3,4 @@ import streamlit.web.bootstrap as bootstrap
 
 if __name__ == '__main__':
     st._is_running_with_streamlit = True
-    bootstrap.run('Waters2mzML-ui.py', 'streamlit run', args=[], flag_options={})
+    bootstrap.run('Select_raw_folders.py', 'streamlit run', args=[], flag_options={})
